now supports DNS update; reorganized code
