+++ src/cli/lib/handle-update.mjs
@@ -1,23 +1,21 @@
 import commandLineArgs from 'command-line-args'
 
 import { cliSpec } from '../constants'
+import { getSiteInfo } from './get-site-info'
 import { update } from '../../lib/actions/update'
 
 const handleUpdate = async ({ argv, globalOptions, sitesInfo }) => {
   const updateOptionsSpec = cliSpec.commands.find(({ name }) => name === 'update').arguments
   const updateOptions = commandLineArgs(updateOptionsSpec, { argv })
   const apexDomain = updateOptions['apex-domain']
+  const doContent = updateOptions['do-content']
+  const doDNS = updateOptions['do-dns']
   const noBuild = updateOptions['no-build']
   const noCacheInvalidation = updateOptions['no-cache-invalidation']
-  const onlyContent = updateOptions['only-content']
 
-  const siteInfo = sitesInfo[apexDomain]
-  if (siteInfo === undefined) {
-    process.stderr.write(`No such site '${apexDomain}' found.\n`)
-    process.exit(1) // eslint-disable-line no-process-exit
-  }
+  const siteInfo = getSiteInfo({ apexDomain, sitesInfo })
 
-  await update({ noBuild, noCacheInvalidation, onlyContent, siteInfo, ...globalOptions })
+  await update({ doContent, doDNS, noBuild, noCacheInvalidation, siteInfo, ...globalOptions })
 }
 
 export { handleUpdate }
